Only return `net_claimable_consensus_rewards` as `value` before withdrawal (#1963)
When exit requests are being proposed, the net claimable rewards theoretically still exists on Kiln's API. It is therefore returned in the confirmation view entity. After execution, it is no longer there. Therefore, it has been removed from the `txInfo` as the value would disappear when moving to the history. (Web will display number of validators instead.)

Withdrawal requests now fetch the net claimable rewards when queued but instead decode the value from the relevant `Withdrawal` events of an executed withdrawal:

- Remove `value` from `NativeStakingValidatorsExitTransactionInfo`.
- Implement `Withdrawal` event decoding.
- Move methods shared between transaction/confirmation view mapping to `KilnNativeStakingHelper`.
- Add relative `Withdrawal` encoder/builder.
- Decode `value` of withdrawals after execution.
- Add/update tests accordingly.